feat: Update Frida tools and agent files
This commit updates the Frida tools and agent files to adapt to the `frida:rpc` feature modification. It solves the problem of scanning a specific string in Android memory and automatically scans the location of the locally installed pip Frida tools. The `core.py` file and the `_frida.abi3.so` file are modified accordingly. Additionally, the frida-agent.dylib file is hidden by changing its file name and loading location. These changes enhance the functionality and security of the project.